REFACTOR: app.py utils.py main.py 로 코드 분리
