+++ main.py
@@ -1,0 +1,7 @@
+from app import run_app
+
+def main():
+    run_app()
+
+if __name__ == "__main__":
+    main()
